+++ client/app/components/groupDetail/store/storeDetail.controller.js
@@ -6,7 +6,7 @@ class StoreDetailController {
       Geocoding,
       $scope
     });
-    
+
     this.pickupListOptions = {
       showCreateButton: true,
       filter: {
