+++ client/app/components/groupDetail/store/storeDetail.html
@@ -1,49 +1,55 @@
-<div class="storeDetail layout-padding">
+<div class="storeDetail">
   <form editable-form name="editableStore" onaftersave="$ctrl.updateStoredata()">
-    <div layout="row">
-      <h1 class="md-headline" flex editable-text="$ctrl.storedata.name">
-        <i class="fa fa-shopping-cart" aria-hidden="true"></i>
-        {{$ctrl.storedata.name}}
-      </h1>
-      <md-button class="md-icon-button" ng-click="editableStore.$show()" ng-show="!editableStore.$visible"
-        aria-label="{{ 'STOREDETAIL.EDIT' | translate }}">
-        <md-tooltip md-direction="left">
-          <span translate="STOREDETAIL.EDIT"></span>
-        </md-tooltip>
-        <i class="fa fa-pencil"></i>
-      </md-button>
-    </div>
-    <div ng-hide="editableStore.$visible">
-      <p ng-if="$ctrl.storedata.address">
-        <md-tooltip>
-          <span translate="STOREDETAIL.ADDRESS"></span>
-        </md-tooltip>
-        <i class="fa fa-map-marker"></i>
-        {{$ctrl.storedata.address}}
-      </p>
-    </div>
-    <div ng-if="editableStore.$visible">
-      <md-input-container>
-        <label translate="STOREDETAIL.ADDRESS"></label>
-        <input type="text" ng-model="$ctrl.storeEdit.address">
-      </md-input-container>
-      <md-button ng-click="$ctrl.addressLookup()" class="md-icon-button" ng-disabled="!$ctrl.storeEdit.address || $ctrl.lookupOngoing">
-        <i class="fa fa-search" ng-show="!$ctrl.lookupOngoing"></i>
-        <i class="fa fa-spinner" ng-show="$ctrl.lookupOngoing"></i>
-      </md-button>
-    </div>
-    <div editable-textarea="$ctrl.storedata.description">
-      <p class="multiline">{{$ctrl.storedata.description}}</p>
-    </div>
-    <div ng-if="editableStore.$visible" layout="row" layout-align="end start">
-      <md-button class="md-icon-button" ng-disabled="editableStore.$waiting" ng-click="$ctrl.reset()">
-        <i class="fa fa-times"></i>
-      </md-button>
-      <md-button type="submit" class="md-raised md-primary" ng-disabled="editableStore.$waiting">
-        <i class="fa fa-check"></i>
-      </md-button>
-    </div>
-  </form>
-  <store-detail-map ng-if="$ctrl.mapData.latitude && $ctrl.mapData.longitude" store-data="$ctrl.mapData"></store-detail-map>
-  <pickup-list store-id="$ctrl.storedata.id" options="$ctrl.pickupListOptions"></pickup-list>
+    <md-toolbar class="md-hue-1">
+      <div class="md-toolbar-tools">
+        <h1 class="md-headline" flex editable-text="$ctrl.storedata.name">
+          <i class="fa fa-shopping-cart" aria-hidden="true"></i>
+          {{$ctrl.storedata.name}}
+        </h1>
+        <md-button class="md-icon-button" ng-click="editableStore.$show()" ng-show="!editableStore.$visible"
+          aria-label="{{ 'STOREDETAIL.EDIT' | translate }}">
+          <md-tooltip md-direction="left">
+            <span translate="STOREDETAIL.EDIT"></span>
+          </md-tooltip>
+          <i class="fa fa-pencil"></i>
+        </md-button>
+      </div>
+    </md-toolbar>
+    
+    <md-content layout-padding>
+      <store-detail-map ng-if="$ctrl.mapData.latitude && $ctrl.mapData.longitude" store-data="$ctrl.mapData"></store-detail-map>
+      
+      <div ng-hide="editableStore.$visible">
+        <p ng-if="$ctrl.storedata.address">
+          <md-tooltip>
+            <span translate="STOREDETAIL.ADDRESS"></span>
+          </md-tooltip>
+          <i class="fa fa-map-marker"></i>
+          {{$ctrl.storedata.address}}
+        </p>
+      </div>
+      <div ng-if="editableStore.$visible">
+        <md-input-container>
+          <label translate="STOREDETAIL.ADDRESS"></label>
+          <input type="text" ng-model="$ctrl.storeEdit.address">
+        </md-input-container>
+        <md-button ng-click="$ctrl.addressLookup()" class="md-icon-button" ng-disabled="!$ctrl.storeEdit.address || $ctrl.lookupOngoing">
+          <i class="fa fa-search" ng-show="!$ctrl.lookupOngoing"></i>
+          <i class="fa fa-spinner" ng-show="$ctrl.lookupOngoing"></i>
+        </md-button>
+      </div>
+      <div editable-textarea="$ctrl.storedata.description">
+        <p class="multiline">{{$ctrl.storedata.description}}</p>
+      </div>
+      <div ng-if="editableStore.$visible" layout="row" layout-align="end start">
+        <md-button class="md-icon-button" ng-disabled="editableStore.$waiting" ng-click="$ctrl.reset()">
+          <i class="fa fa-times"></i>
+        </md-button>
+        <md-button type="submit" class="md-raised md-primary" ng-disabled="editableStore.$waiting">
+          <i class="fa fa-check"></i>
+        </md-button>
+      </div>
+    </form>
+    <pickup-list store-id="$ctrl.storedata.id" options="$ctrl.pickupListOptions"></pickup-list>
+  </md-content>
 </div>
